Update HMock/explainable-predicates config (Cabal)
A Hackage revision was made that fixes the upper bounds for both `base`
and `template-haskell` for these packages.  Cabal no longer needs the
`allow-newer` configuration.

Note that removing the `allow-newer` configuration in Stack still
results in failure.  I suspect that pre-revision constraints are still
cached somewhere and plan on trying again after waiting a bit.